Adjusted to Black formatting

+++ utils.py
@@ -1,17 +1,22 @@
 """Simple math utility functions."""
 
-def add (a: int, b: int) -> int:
+
+def add(a: int, b: int) -> int:
     """Returns the sum of a and b."""
     return a + b
 
-def substract (a: int, b: int) -> int:
+
+def substract(a: int, b: int) -> int:
     """Returns the substraction result of a and b."""
     return a - b
+
 
 def multiply(a: int, b: int) -> int:
     """Returns the product of a and b."""
     return a * b
 
-def divide(a: int, b: int) -> int:
+
+def divide(a: int, b: int) -> float:
     """Returns the quotient of a and b."""
     return a / b
+
